☑️Fixed drawpile orientation, added a drawing state to gameplay, refresh button, helpful gameplay messages, and textbox functionality to lobby options
Drawpile just needed an h and v flip to look right.

I added a drawing state which requires the player to chose either the drawpile or discardpile to draw a card from before going into the discard state.

I added messages in the play area that appear after a set amount of time to help a player who is unaware what theyre options are.

I added a refresh button to the lobbylist page to refresh available lobbies.

Finally I copied and edited the code from the join by ID promt to give an area where a password can be set and a point goal can be pick out.

While I was at it, I added the button overlay background so that you can click out of the textbox prompt when asked to set a password or lobbyID etc.